Django Movie_MiniPJT RESTful

+++ Web/04_django/Open_API/DJANGO_BLOG/movies/views.py
@@ -9,42 +9,51 @@ def new(request):
     return render(request, 'movies/new.html')
 
 def create(request):
-    movie = Movie()
-    movie.title = request.GET.get('title')
-    movie.title_en = request.GET.get('title_en')
-    movie.audience = request.GET.get('audience')
-    movie.open_date = request.GET.get('open_date')
-    movie.genre = request.GET.get('genre')
-    movie.watch_grade = request.GET.get('watch_grade')
-    movie.score = request.GET.get('score')
-    movie.poster_url = request.GET.get('poster_url')
-    movie.description = request.GET.get('description')
-    movie.save()
-    return render(request, 'movies/create.html', {'movie':movie})
+    if request.method == 'POST':
+        movie = Movie()
+        movie.title = request.POST.get('title')
+        movie.title_en = request.POST.get('title_en')
+        movie.audience = request.POST.get('audience')
+        movie.open_date = request.POST.get('open_date')
+        movie.genre = request.POST.get('genre')
+        movie.watch_grade = request.POST.get('watch_grade')
+        movie.score = request.POST.get('score')
+        movie.poster_url = request.POST.get('poster_url')
+        movie.description = request.POST.get('description')
+        movie.save()
+        return redirect('movies:detail', movie.pk)
+    else:
+        return render(request, 'movies/new.html')
 
 def detail(request, pk):
     movie = Movie.objects.get(pk=pk)
     return render(request, 'movies/detail.html', {'movie':movie})
 
-def edit(request, pk):
-    movie = Movie.objects.get(pk=pk)
-    return render(request, 'movies/edit.html', {'movie':movie})
+# def edit(request, pk):
+#     movie = Movie.objects.get(pk=pk)
+#     return render(request, 'movies/edit.html', {'movie':movie})
 
 def update(request, pk):
     movie = Movie.objects.get(pk=pk)
-    movie.title = request.GET.get('title')
-    movie.title_en = request.GET.get('title_en')
-    movie.audience = request.GET.get('audience')
-    movie.open_date = request.GET.get('open_date')
-    movie.genre = request.GET.get('genre')
-    movie.watch_grade = request.GET.get('watch_grade')
-    movie.score = request.GET.get('score')
-    movie.poster_url = request.GET.get('poster_url')
-    movie.description = request.GET.get('description')
-    movie.save()
-    return render(request, 'movies/update.html', {'movie':movie})
+    if request.method == 'POST':
+        movie.title = request.POST.get('title')
+        movie.title_en = request.POST.get('title_en')
+        movie.audience = request.POST.get('audience')
+        movie.open_date = request.POST.get('open_date')
+        movie.genre = request.POST.get('genre')
+        movie.watch_grade = request.POST.get('watch_grade')
+        movie.score = request.POST.get('score')
+        movie.poster_url = request.POST.get('poster_url')
+        movie.description = request.POST.get('description')
+        movie.save()
+        return redirect('movies:detail', movie.pk)
+    else:
+        return render(request, 'movies/edit.html', {'movie':movie})
 
 def delete(request, pk):
     movie = Movie.objects.get(pk=pk)
-    movie.delete()
-    return redirect('/movie/')
+    if request.method == 'POST':
+        movie.delete()
+        return redirect('movies:index')
+    else:
+        return redirect('movies:detail', movie.pk)
